tests/localstack/oidc/requirements.txt
flask==3.0.0
pyjwt==2.8.0
cryptography==42.0.0
